ref:/ change structure with beter practis
